Move conceptual models to core
